refactor(metrics): split aggregate tests for GH727
